refactor: event gateway type 파일 분리
